FLASK/flask-authz01/requirements.txt
casbin==0.8.3
Click==7.0
Flask==1.1.1
flask-authz==0.0.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
simpleeval==0.9.10
Werkzeug==0.16.0
